Move database test artifacts to database module
