+++ taskodoro/build.gradle.kts
@@ -1,5 +1,5 @@
 /*
- *    Copyright 2023 Felipe Joglar
+ *    Copyright 2024 Felipe Joglar
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -61,7 +61,6 @@ kotlin {
         /* Test source sets */
         val commonTest by getting {
             dependencies {
-                implementation(projects.infra.databaseTest)
                 implementation(libs.kotlin.test)
             }
         }
